Fix documentation reg. unittest, add run_tests.py
There is a problem with running nosetest - the previous html coverage reports
are not properly removed thus subsequent runs of nosetest --with-coverage...
result in "File exists error".

The old run_tests.py file was restored and put into the main directory. It takes
care of removing old html reports and running the tests.